Add permissions for all views

+++ main/permissions.py
@@ -26,20 +26,25 @@ def validate_custom_token(request):
     if len(token) == 16:
         token = AdminToken.objects.filter(key=token)
         if token and token.first().expiration_time < timezone.now() - \
-            timedelta(hours=0.01):
+            timedelta(hours=1):
             token.delete()
             raise AuthenticationFailed('Токен истек')
         elif token:
             return True if token.first().user.is_superuser else False
         else:
             raise AuthenticationFailed('Токена не существует, создайте новый')
+        
     elif len(token) == 32:
         token = UserToken.objects.filter(key=token)
-        if token and token.first().expiration_time < timezone.now() - \
-            timedelta(hours=1):
-            token.delete()
-            raise AuthenticationFailed('Токен истек')
-        elif token:
-            return True if token else False
+        if request.method == 'GET':
+            if token and token.first().expiration_time < timezone.now() - \
+                timedelta(hours=1):
+                token.delete()
+                raise AuthenticationFailed('Токен истек')
+            elif token:
+                return True if token else False
+            else:
+                raise AuthenticationFailed\
+                    ('Токен не существует, создайте новый')
         else:
-            raise AuthenticationFailed('Токена не существует, создайте новый')
+            raise AuthenticationFailed('Доступ запрещен')
